:recycle: Move EmptyState to toolbox/box
