+++ src/components/toolbox/box/emptyState.css
@@ -1,5 +1,5 @@
-@import '../app/variables.css';
-@import '../app/mixins.css';
+@import '../../app/variables.css';
+@import '../../app/mixins.css';
 
 .wrapper {
   width: 100%;
