+++ src/components/toolbox/box/index.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import Content from './content';
-import EmptyState from '../../emptyState';
+import EmptyState from './emptyState';
 import Footer from './footer';
 import FooterButton from './footerButton';
 import Header from './header';
